[WIP COMMIT] preparing to migrate cn
This commit removes all non-cn files and adapts the build scripts to use
cerberus as a library.

DO NOT WORK ON TOP OF THIS COMMIT.

This is meant to be the starting point of the new rems-project/cn repository.

I will keep on force pushing until this commit is finalised.